chane type column Code

+++ write-api/utils/res/response_http.go
@@ -5,7 +5,7 @@ import "time"
 type ResponseHttp[T any] struct {
 	Timestamp time.Time `json:"timestamp"`
 	Payload   T         `json:"payload"`
-	Code      int16     `json:"code"`
+	Code      int       `json:"code"`
 	Status    bool      `json:"status"`
 	Message   string    `json:"message"`
 	Version   uint8     `json:"version"`
